+++ cmd/zz_gen_cmd_dnshelp.go
@@ -81,7 +81,7 @@ func displayDNSHelp(name string) {
 	switch name {
 
 	case "acme-dns":
-		// generated from: providers\dns\acmedns\acmedns.toml
+		// generated from: providers/dns/acmedns/acmedns.toml
 		fmt.Fprintln(w, `Configuration for Joohoi's ACME-DNS.`)
 		fmt.Fprintln(w, `Code: 'acme-dns'`)
 		fmt.Fprintln(w, `Since: 'v1.1.0'`)
@@ -96,7 +96,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/acme-dns`)
 
 	case "alidns":
-		// generated from: providers\dns\alidns\alidns.toml
+		// generated from: providers/dns/alidns/alidns.toml
 		fmt.Fprintln(w, `Configuration for Alibaba Cloud DNS.`)
 		fmt.Fprintln(w, `Code: 'alidns'`)
 		fmt.Fprintln(w, `Since: 'v1.1.0'`)
@@ -117,7 +117,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/alidns`)
 
 	case "auroradns":
-		// generated from: providers\dns\auroradns\auroradns.toml
+		// generated from: providers/dns/auroradns/auroradns.toml
 		fmt.Fprintln(w, `Configuration for Aurora DNS.`)
 		fmt.Fprintln(w, `Code: 'auroradns'`)
 		fmt.Fprintln(w, `Since: 'v0.4.0'`)
@@ -138,7 +138,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/auroradns`)
 
 	case "azure":
-		// generated from: providers\dns\azure\azure.toml
+		// generated from: providers/dns/azure/azure.toml
 		fmt.Fprintln(w, `Configuration for Azure.`)
 		fmt.Fprintln(w, `Code: 'azure'`)
 		fmt.Fprintln(w, `Since: 'v0.4.0'`)
@@ -163,7 +163,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/azure`)
 
 	case "bluecat":
-		// generated from: providers\dns\bluecat\bluecat.toml
+		// generated from: providers/dns/bluecat/bluecat.toml
 		fmt.Fprintln(w, `Configuration for Bluecat.`)
 		fmt.Fprintln(w, `Code: 'bluecat'`)
 		fmt.Fprintln(w, `Since: 'v0.5.0'`)
@@ -187,7 +187,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/bluecat`)
 
 	case "cloudflare":
-		// generated from: providers\dns\cloudflare\cloudflare.toml
+		// generated from: providers/dns/cloudflare/cloudflare.toml
 		fmt.Fprintln(w, `Configuration for Cloudflare.`)
 		fmt.Fprintln(w, `Code: 'cloudflare'`)
 		fmt.Fprintln(w, `Since: 'v0.3.0'`)
@@ -208,7 +208,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/cloudflare`)
 
 	case "cloudns":
-		// generated from: providers\dns\cloudns\cloudns.toml
+		// generated from: providers/dns/cloudns/cloudns.toml
 		fmt.Fprintln(w, `Configuration for ClouDNS.`)
 		fmt.Fprintln(w, `Code: 'cloudns'`)
 		fmt.Fprintln(w, `Since: 'v2.3.0'`)
@@ -229,7 +229,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/cloudns`)
 
 	case "cloudxns":
-		// generated from: providers\dns\cloudxns\cloudxns.toml
+		// generated from: providers/dns/cloudxns/cloudxns.toml
 		fmt.Fprintln(w, `Configuration for CloudXNS.`)
 		fmt.Fprintln(w, `Code: 'cloudxns'`)
 		fmt.Fprintln(w, `Since: 'v0.5.0'`)
@@ -250,7 +250,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/cloudxns`)
 
 	case "conoha":
-		// generated from: providers\dns\conoha\conoha.toml
+		// generated from: providers/dns/conoha/conoha.toml
 		fmt.Fprintln(w, `Configuration for ConoHa.`)
 		fmt.Fprintln(w, `Code: 'conoha'`)
 		fmt.Fprintln(w, `Since: 'v1.2.0'`)
@@ -273,7 +273,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/conoha`)
 
 	case "designate":
-		// generated from: providers\dns\designate\designate.toml
+		// generated from: providers/dns/designate/designate.toml
 		fmt.Fprintln(w, `Configuration for Designate DNSaaS for Openstack.`)
 		fmt.Fprintln(w, `Code: 'designate'`)
 		fmt.Fprintln(w, `Since: 'v2.2.0'`)
@@ -296,7 +296,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/designate`)
 
 	case "digitalocean":
-		// generated from: providers\dns\digitalocean\digitalocean.toml
+		// generated from: providers/dns/digitalocean/digitalocean.toml
 		fmt.Fprintln(w, `Configuration for Digital Ocean.`)
 		fmt.Fprintln(w, `Code: 'digitalocean'`)
 		fmt.Fprintln(w, `Since: 'v0.3.0'`)
@@ -316,7 +316,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/digitalocean`)
 
 	case "dnsimple":
-		// generated from: providers\dns\dnsimple\dnsimple.toml
+		// generated from: providers/dns/dnsimple/dnsimple.toml
 		fmt.Fprintln(w, `Configuration for DNSimple.`)
 		fmt.Fprintln(w, `Code: 'dnsimple'`)
 		fmt.Fprintln(w, `Since: 'v0.3.0'`)
@@ -336,7 +336,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/dnsimple`)
 
 	case "dnsmadeeasy":
-		// generated from: providers\dns\dnsmadeeasy\dnsmadeeasy.toml
+		// generated from: providers/dns/dnsmadeeasy/dnsmadeeasy.toml
 		fmt.Fprintln(w, `Configuration for DNS Made Easy.`)
 		fmt.Fprintln(w, `Code: 'dnsmadeeasy'`)
 		fmt.Fprintln(w, `Since: 'v0.4.0'`)
@@ -358,7 +358,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/dnsmadeeasy`)
 
 	case "dnspod":
-		// generated from: providers\dns\dnspod\dnspod.toml
+		// generated from: providers/dns/dnspod/dnspod.toml
 		fmt.Fprintln(w, `Configuration for DNSPod.`)
 		fmt.Fprintln(w, `Code: 'dnspod'`)
 		fmt.Fprintln(w, `Since: 'v0.4.0'`)
@@ -378,7 +378,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/dnspod`)
 
 	case "dode":
-		// generated from: providers\dns\dode\dode.toml
+		// generated from: providers/dns/dode/dode.toml
 		fmt.Fprintln(w, `Configuration for Domain Offensive (do.de).`)
 		fmt.Fprintln(w, `Code: 'dode'`)
 		fmt.Fprintln(w, `Since: 'v2.4.0'`)
@@ -399,7 +399,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/dode`)
 
 	case "dreamhost":
-		// generated from: providers\dns\dreamhost\dreamhost.toml
+		// generated from: providers/dns/dreamhost/dreamhost.toml
 		fmt.Fprintln(w, `Configuration for DreamHost.`)
 		fmt.Fprintln(w, `Code: 'dreamhost'`)
 		fmt.Fprintln(w, `Since: 'v1.1.0'`)
@@ -419,7 +419,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/dreamhost`)
 
 	case "duckdns":
-		// generated from: providers\dns\duckdns\duckdns.toml
+		// generated from: providers/dns/duckdns/duckdns.toml
 		fmt.Fprintln(w, `Configuration for Duck DNS.`)
 		fmt.Fprintln(w, `Code: 'duckdns'`)
 		fmt.Fprintln(w, `Since: 'v0.5.0'`)
@@ -440,7 +440,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/duckdns`)
 
 	case "dyn":
-		// generated from: providers\dns\dyn\dyn.toml
+		// generated from: providers/dns/dyn/dyn.toml
 		fmt.Fprintln(w, `Configuration for Dyn.`)
 		fmt.Fprintln(w, `Code: 'dyn'`)
 		fmt.Fprintln(w, `Since: 'v0.3.0'`)
@@ -462,10 +462,10 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/dyn`)
 
 	case "easydns":
-		// generated from: providers\dns\easydns\easydns.toml
+		// generated from: providers/dns/easydns/easydns.toml
 		fmt.Fprintln(w, `Configuration for EasyDNS.`)
 		fmt.Fprintln(w, `Code: 'easydns'`)
-		fmt.Fprintln(w, `Since: ''`)
+		fmt.Fprintln(w, `Since: 'v2.6.0'`)
 		fmt.Fprintln(w)
 
 		fmt.Fprintln(w, `Credentials:`)
@@ -485,7 +485,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/easydns`)
 
 	case "exec":
-		// generated from: providers\dns\exec\exec.toml
+		// generated from: providers/dns/exec/exec.toml
 		fmt.Fprintln(w, `Configuration for External program.`)
 		fmt.Fprintln(w, `Code: 'exec'`)
 		fmt.Fprintln(w, `Since: 'v0.5.0'`)
@@ -495,7 +495,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/exec`)
 
 	case "exoscale":
-		// generated from: providers\dns\exoscale\exoscale.toml
+		// generated from: providers/dns/exoscale/exoscale.toml
 		fmt.Fprintln(w, `Configuration for Exoscale.`)
 		fmt.Fprintln(w, `Code: 'exoscale'`)
 		fmt.Fprintln(w, `Since: 'v0.4.0'`)
@@ -517,7 +517,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/exoscale`)
 
 	case "fastdns":
-		// generated from: providers\dns\fastdns\fastdns.toml
+		// generated from: providers/dns/fastdns/fastdns.toml
 		fmt.Fprintln(w, `Configuration for FastDNS.`)
 		fmt.Fprintln(w, `Code: 'fastdns'`)
 		fmt.Fprintln(w, `Since: 'v0.5.0'`)
@@ -539,7 +539,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/fastdns`)
 
 	case "gandi":
-		// generated from: providers\dns\gandi\gandi.toml
+		// generated from: providers/dns/gandi/gandi.toml
 		fmt.Fprintln(w, `Configuration for Gandi.`)
 		fmt.Fprintln(w, `Code: 'gandi'`)
 		fmt.Fprintln(w, `Since: 'v0.3.0'`)
@@ -559,7 +559,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/gandi`)
 
 	case "gandiv5":
-		// generated from: providers\dns\gandiv5\gandiv5.toml
+		// generated from: providers/dns/gandiv5/gandiv5.toml
 		fmt.Fprintln(w, `Configuration for Gandi Live DNS (v5).`)
 		fmt.Fprintln(w, `Code: 'gandiv5'`)
 		fmt.Fprintln(w, `Since: 'v0.5.0'`)
@@ -579,7 +579,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/gandiv5`)
 
 	case "gcloud":
-		// generated from: providers\dns\gcloud\gcloud.toml
+		// generated from: providers/dns/gcloud/gcloud.toml
 		fmt.Fprintln(w, `Configuration for Google Cloud.`)
 		fmt.Fprintln(w, `Code: 'gcloud'`)
 		fmt.Fprintln(w, `Since: 'v0.3.0'`)
@@ -601,7 +601,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/gcloud`)
 
 	case "glesys":
-		// generated from: providers\dns\glesys\glesys.toml
+		// generated from: providers/dns/glesys/glesys.toml
 		fmt.Fprintln(w, `Configuration for Glesys.`)
 		fmt.Fprintln(w, `Code: 'glesys'`)
 		fmt.Fprintln(w, `Since: 'v0.5.0'`)
@@ -622,7 +622,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/glesys`)
 
 	case "godaddy":
-		// generated from: providers\dns\godaddy\godaddy.toml
+		// generated from: providers/dns/godaddy/godaddy.toml
 		fmt.Fprintln(w, `Configuration for Go Daddy.`)
 		fmt.Fprintln(w, `Code: 'godaddy'`)
 		fmt.Fprintln(w, `Since: 'v0.5.0'`)
@@ -644,7 +644,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/godaddy`)
 
 	case "hostingde":
-		// generated from: providers\dns\hostingde\hostingde.toml
+		// generated from: providers/dns/hostingde/hostingde.toml
 		fmt.Fprintln(w, `Configuration for Hosting.de.`)
 		fmt.Fprintln(w, `Code: 'hostingde'`)
 		fmt.Fprintln(w, `Since: 'v1.1.0'`)
@@ -665,7 +665,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/hostingde`)
 
 	case "httpreq":
-		// generated from: providers\dns\httpreq\httpreq.toml
+		// generated from: providers/dns/httpreq/httpreq.toml
 		fmt.Fprintln(w, `Configuration for HTTP request.`)
 		fmt.Fprintln(w, `Code: 'httpreq'`)
 		fmt.Fprintln(w, `Since: 'v2.0.0'`)
@@ -687,7 +687,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/httpreq`)
 
 	case "iij":
-		// generated from: providers\dns\iij\iij.toml
+		// generated from: providers/dns/iij/iij.toml
 		fmt.Fprintln(w, `Configuration for Internet Initiative Japan.`)
 		fmt.Fprintln(w, `Code: 'iij'`)
 		fmt.Fprintln(w, `Since: 'v1.1.0'`)
@@ -708,7 +708,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/iij`)
 
 	case "inwx":
-		// generated from: providers\dns\inwx\inwx.toml
+		// generated from: providers/dns/inwx/inwx.toml
 		fmt.Fprintln(w, `Configuration for INWX.`)
 		fmt.Fprintln(w, `Code: 'inwx'`)
 		fmt.Fprintln(w, `Since: 'v2.0.0'`)
@@ -729,7 +729,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/inwx`)
 
 	case "lightsail":
-		// generated from: providers\dns\lightsail\lightsail.toml
+		// generated from: providers/dns/lightsail/lightsail.toml
 		fmt.Fprintln(w, `Configuration for Amazon Lightsail.`)
 		fmt.Fprintln(w, `Code: 'lightsail'`)
 		fmt.Fprintln(w, `Since: 'v0.5.0'`)
@@ -749,7 +749,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/lightsail`)
 
 	case "linode":
-		// generated from: providers\dns\linode\linode.toml
+		// generated from: providers/dns/linode/linode.toml
 		fmt.Fprintln(w, `Configuration for Linode (deprecated).`)
 		fmt.Fprintln(w, `Code: 'linode'`)
 		fmt.Fprintln(w, `Since: 'v0.4.0'`)
@@ -768,7 +768,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/linode`)
 
 	case "linodev4":
-		// generated from: providers\dns\linodev4\linodev4.toml
+		// generated from: providers/dns/linodev4/linodev4.toml
 		fmt.Fprintln(w, `Configuration for Linode (v4).`)
 		fmt.Fprintln(w, `Code: 'linodev4'`)
 		fmt.Fprintln(w, `Since: 'v1.1.0'`)
@@ -787,7 +787,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/linodev4`)
 
 	case "mydnsjp":
-		// generated from: providers\dns\mydnsjp\mydnsjp.toml
+		// generated from: providers/dns/mydnsjp/mydnsjp.toml
 		fmt.Fprintln(w, `Configuration for MyDNS.jp.`)
 		fmt.Fprintln(w, `Code: 'mydnsjp'`)
 		fmt.Fprintln(w, `Since: 'v1.2.0'`)
@@ -808,7 +808,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/mydnsjp`)
 
 	case "namecheap":
-		// generated from: providers\dns\namecheap\namecheap.toml
+		// generated from: providers/dns/namecheap/namecheap.toml
 		fmt.Fprintln(w, `Configuration for Namecheap.`)
 		fmt.Fprintln(w, `Code: 'namecheap'`)
 		fmt.Fprintln(w, `Since: 'v0.3.0'`)
@@ -829,7 +829,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/namecheap`)
 
 	case "namedotcom":
-		// generated from: providers\dns\namedotcom\namedotcom.toml
+		// generated from: providers/dns/namedotcom/namedotcom.toml
 		fmt.Fprintln(w, `Configuration for Name.com.`)
 		fmt.Fprintln(w, `Code: 'namedotcom'`)
 		fmt.Fprintln(w, `Since: 'v0.5.0'`)
@@ -850,7 +850,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/namedotcom`)
 
 	case "netcup":
-		// generated from: providers\dns\netcup\netcup.toml
+		// generated from: providers/dns/netcup/netcup.toml
 		fmt.Fprintln(w, `Configuration for Netcup.`)
 		fmt.Fprintln(w, `Code: 'netcup'`)
 		fmt.Fprintln(w, `Since: 'v1.1.0'`)
@@ -872,7 +872,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/netcup`)
 
 	case "nifcloud":
-		// generated from: providers\dns\nifcloud\nifcloud.toml
+		// generated from: providers/dns/nifcloud/nifcloud.toml
 		fmt.Fprintln(w, `Configuration for NIFCloud.`)
 		fmt.Fprintln(w, `Code: 'nifcloud'`)
 		fmt.Fprintln(w, `Since: 'v1.1.0'`)
@@ -893,7 +893,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/nifcloud`)
 
 	case "ns1":
-		// generated from: providers\dns\ns1\ns1.toml
+		// generated from: providers/dns/ns1/ns1.toml
 		fmt.Fprintln(w, `Configuration for NS1.`)
 		fmt.Fprintln(w, `Code: 'ns1'`)
 		fmt.Fprintln(w, `Since: 'v0.4.0'`)
@@ -913,7 +913,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/ns1`)
 
 	case "oraclecloud":
-		// generated from: providers\dns\oraclecloud\oraclecloud.toml
+		// generated from: providers/dns/oraclecloud/oraclecloud.toml
 		fmt.Fprintln(w, `Configuration for Oracle Cloud.`)
 		fmt.Fprintln(w, `Code: 'oraclecloud'`)
 		fmt.Fprintln(w, `Since: 'v2.3.0'`)
@@ -938,7 +938,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/oraclecloud`)
 
 	case "otc":
-		// generated from: providers\dns\otc\otc.toml
+		// generated from: providers/dns/otc/otc.toml
 		fmt.Fprintln(w, `Configuration for Open Telekom Cloud.`)
 		fmt.Fprintln(w, `Code: 'otc'`)
 		fmt.Fprintln(w, `Since: 'v0.4.1'`)
@@ -962,7 +962,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/otc`)
 
 	case "ovh":
-		// generated from: providers\dns\ovh\ovh.toml
+		// generated from: providers/dns/ovh/ovh.toml
 		fmt.Fprintln(w, `Configuration for OVH.`)
 		fmt.Fprintln(w, `Code: 'ovh'`)
 		fmt.Fprintln(w, `Since: 'v0.4.0'`)
@@ -985,7 +985,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/ovh`)
 
 	case "pdns":
-		// generated from: providers\dns\pdns\pdns.toml
+		// generated from: providers/dns/pdns/pdns.toml
 		fmt.Fprintln(w, `Configuration for PowerDNS.`)
 		fmt.Fprintln(w, `Code: 'pdns'`)
 		fmt.Fprintln(w, `Since: 'v0.4.0'`)
@@ -1006,7 +1006,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/pdns`)
 
 	case "rackspace":
-		// generated from: providers\dns\rackspace\rackspace.toml
+		// generated from: providers/dns/rackspace/rackspace.toml
 		fmt.Fprintln(w, `Configuration for Rackspace.`)
 		fmt.Fprintln(w, `Code: 'rackspace'`)
 		fmt.Fprintln(w, `Since: 'v0.4.0'`)
@@ -1027,7 +1027,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/rackspace`)
 
 	case "rfc2136":
-		// generated from: providers\dns\rfc2136\rfc2136.toml
+		// generated from: providers/dns/rfc2136/rfc2136.toml
 		fmt.Fprintln(w, `Configuration for RFC2136.`)
 		fmt.Fprintln(w, `Code: 'rfc2136'`)
 		fmt.Fprintln(w, `Since: 'v0.3.0'`)
@@ -1051,7 +1051,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/rfc2136`)
 
 	case "route53":
-		// generated from: providers\dns\route53\route53.toml
+		// generated from: providers/dns/route53/route53.toml
 		fmt.Fprintln(w, `Configuration for Amazon Route 53.`)
 		fmt.Fprintln(w, `Code: 'route53'`)
 		fmt.Fprintln(w, `Since: 'v0.3.0'`)
@@ -1074,7 +1074,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/route53`)
 
 	case "sakuracloud":
-		// generated from: providers\dns\sakuracloud\sakuracloud.toml
+		// generated from: providers/dns/sakuracloud/sakuracloud.toml
 		fmt.Fprintln(w, `Configuration for Sakura Cloud.`)
 		fmt.Fprintln(w, `Code: 'sakuracloud'`)
 		fmt.Fprintln(w, `Since: 'v1.1.0'`)
@@ -1095,7 +1095,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/sakuracloud`)
 
 	case "selectel":
-		// generated from: providers\dns\selectel\selectel.toml
+		// generated from: providers/dns/selectel/selectel.toml
 		fmt.Fprintln(w, `Configuration for Selectel.`)
 		fmt.Fprintln(w, `Code: 'selectel'`)
 		fmt.Fprintln(w, `Since: 'v1.2.0'`)
@@ -1116,7 +1116,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/selectel`)
 
 	case "stackpath":
-		// generated from: providers\dns\stackpath\stackpath.toml
+		// generated from: providers/dns/stackpath/stackpath.toml
 		fmt.Fprintln(w, `Configuration for Stackpath.`)
 		fmt.Fprintln(w, `Code: 'stackpath'`)
 		fmt.Fprintln(w, `Since: 'v1.1.0'`)
@@ -1137,7 +1137,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/stackpath`)
 
 	case "transip":
-		// generated from: providers\dns\transip\transip.toml
+		// generated from: providers/dns/transip/transip.toml
 		fmt.Fprintln(w, `Configuration for TransIP.`)
 		fmt.Fprintln(w, `Code: 'transip'`)
 		fmt.Fprintln(w, `Since: 'v2.0.0'`)
@@ -1157,7 +1157,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/transip`)
 
 	case "vegadns":
-		// generated from: providers\dns\vegadns\vegadns.toml
+		// generated from: providers/dns/vegadns/vegadns.toml
 		fmt.Fprintln(w, `Configuration for VegaDNS.`)
 		fmt.Fprintln(w, `Code: 'vegadns'`)
 		fmt.Fprintln(w, `Since: 'v1.1.0'`)
@@ -1178,7 +1178,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/vegadns`)
 
 	case "vscale":
-		// generated from: providers\dns\vscale\vscale.toml
+		// generated from: providers/dns/vscale/vscale.toml
 		fmt.Fprintln(w, `Configuration for Vscale.`)
 		fmt.Fprintln(w, `Code: 'vscale'`)
 		fmt.Fprintln(w, `Since: 'v2.0.0'`)
@@ -1199,7 +1199,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/vscale`)
 
 	case "vultr":
-		// generated from: providers\dns\vultr\vultr.toml
+		// generated from: providers/dns/vultr/vultr.toml
 		fmt.Fprintln(w, `Configuration for Vultr.`)
 		fmt.Fprintln(w, `Code: 'vultr'`)
 		fmt.Fprintln(w, `Since: 'v0.3.1'`)
@@ -1219,7 +1219,7 @@ func displayDNSHelp(name string) {
 		fmt.Fprintln(w, `More information: https://go-acme.github.io/lego/dns/vultr`)
 
 	case "zoneee":
-		// generated from: providers\dns\zoneee\zoneee.toml
+		// generated from: providers/dns/zoneee/zoneee.toml
 		fmt.Fprintln(w, `Configuration for Zone.ee.`)
 		fmt.Fprintln(w, `Code: 'zoneee'`)
 		fmt.Fprintln(w, `Since: 'v2.1.0'`)
